refactor: card examples
